+++ aidial_adapter_openai/chat_completions/tokenizer_factory.py
@@ -33,9 +33,7 @@ from aidial_adapter_openai.configuration.deployment_type import (
     ChatCompletionDeploymentType as D,
 )
 from aidial_adapter_openai.dial_api.storage import FileStorage
-from aidial_adapter_openai.responses.converter import (
-    chat_completions_to_responses_request,
-)
+from aidial_adapter_openai.responses.tokenizer import ResponsesRequestTokenizer
 from aidial_adapter_openai.utils.client import get_client
 from aidial_adapter_openai.utils.request import get_api_version
 from aidial_adapter_openai.utils.tokenizer import (
@@ -116,35 +114,6 @@ class _AnthropicRequestTokenizer:
 
     async def tokenize_raw_request(self, request: dict) -> int:
         raise ValueError("Raw request tokenization isn't supported")
-
-
-@dataclass
-class ResponsesRequestTokenizer:
-    client: AsyncOpenAI
-    file_storage: FileStorage | None
-
-    async def tokenize_text(self, model_name: str, text: str) -> int:
-        response = await self.client.responses.input_tokens.count(
-            model=model_name, input=text
-        )
-        return response.input_tokens
-
-    async def tokenize_request(self, request: ChatCompletionRequest) -> int:
-        return await self.tokenize_raw_request(
-            request.model_dump(exclude_none=True)
-        )
-
-    async def tokenize_raw_request(self, request: dict) -> int:
-        tokenize_request, _ = await chat_completions_to_responses_request(
-            request=request, file_storage=self.file_storage
-        )
-        response = await self.client.responses.input_tokens.count(
-            **tokenize_request
-        )
-        return response.input_tokens
-
-    async def tokenize(self, request: dict) -> int:
-        return await self.tokenize_raw_request(request)
 
 
 async def create_request_tokenizer(
